afm_transformation replaced by antlr parser
